style: rename pkg/apiserver to pkg/server
